webextensions/browser.runtime.html

<!DOCTYPE HTML>
<title>Test browser.runtime</title>
<meta name="extension" content="resources/runtime">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  const asyncTests = {};

  // Test are executed by the extensions inside of their background script
  // or in a content script. The results are sent back to the test runner
  // via postMessage.
  window.addEventListener("message", async ({data}) => {
    // WPTTestResults is a bool set on the test runner being run by
    // the extension.
    if (data.WPTTestResults) {
      const {name} = data

      if (data.isAsyncTest) {
        if (data.asyncResult !== undefined) {
          const t = asyncTests[name]
          t.step(() => {
            assert_true(data.asyncResult, name);
          })
          t.done();
        } else {
          asyncTests[name] = async_test(name);
        }
      } else {
        test(() => {
          assert_true(data.result, name)
        }, /* TODO better titles */ name.replace('Exists', ' Exists'))
      }
    }
  })
</script>
<iframe id="testIframe"></iframe>
